% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{calculate_i2}
\alias{calculate_i2}
\title{Calculate I-squared Statistic}
\usage{
calculate_i2(Q, df)
}
\arguments{
\item{Q}{Cochran's Q statistic}

\item{df}{Degrees of freedom}
}
\value{
I-squared value and confidence intervals
}
\description{
Calculate I-squared Statistic
}
