% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{calculate_tau2}
\alias{calculate_tau2}
\title{Calculate Tau-squared}
\usage{
calculate_tau2(Q, df, weights)
}
\arguments{
\item{Q}{Cochran's Q statistic}

\item{df}{Degrees of freedom}

\item{weights}{Study weights}
}
\value{
Tau-squared estimate
}
\description{
Calculate Tau-squared
}
