% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{cbamm_fast}
\alias{cbamm_fast}
\title{Fast Meta-Analysis}
\usage{
cbamm_fast(data, method = "REML", confidence_level = 0.95)
}
\arguments{
\item{data}{Data frame with effect sizes}

\item{method}{Estimation method ("REML", "DL", "ML")}

\item{confidence_level}{Confidence level for CIs (default: 0.95)}
}
\value{
CBAMM meta-analysis object
}
\description{
Fast Meta-Analysis
}
