% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{compute_analysis_weights}
\alias{compute_analysis_weights}
\title{Compute Analysis Weights}
\usage{
compute_analysis_weights(data, transport_weights = NULL)
}
\arguments{
\item{data}{Study data}

\item{transport_weights}{Transport weights}
}
\value{
Analysis weights
}
\description{
Compute Analysis Weights
}
