% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{compute_transport_weights}
\alias{compute_transport_weights}
\title{Compute Transport Weights for Generalizability}
\usage{
compute_transport_weights(data, target_population, truncation = 0.02)
}
\arguments{
\item{data}{Study data with covariate information}

\item{target_population}{Target population characteristics}

\item{truncation}{Weight truncation level}
}
\value{
Vector of transport weights
}
\description{
Compute Transport Weights for Generalizability
}
