% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{create_cumulative_dashboard}
\alias{create_cumulative_dashboard}
\title{Create Cumulative Analysis Dashboard}
\usage{
create_cumulative_dashboard(x, save_plot = FALSE, filename = "dashboard.png")
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{save_plot}{Whether to save plot}

\item{filename}{Filename for saving}
}
\value{
Dashboard plots
}
\description{
Create Cumulative Analysis Dashboard
}
