% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{detect_outliers_iqr}
\alias{detect_outliers_iqr}
\title{Detect Outliers using IQR Method}
\usage{
detect_outliers_iqr(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Logical vector indicating outliers
}
\description{
Detect Outliers using IQR Method
}
