% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{forest_plot_enhanced}
\alias{forest_plot_enhanced}
\title{Enhanced Forest Plot}
\usage{
forest_plot_enhanced(data, ...)
}
\arguments{
\item{data}{Data frame with study data}

\item{...}{Additional arguments}
}
\value{
ggplot object
}
\description{
Enhanced Forest Plot
}
