% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{meta_regression}
\alias{meta_regression}
\title{Meta-Regression Analysis}
\usage{
meta_regression(data, formula, se = "se")
}
\arguments{
\item{data}{Data frame with meta-analysis data}

\item{formula}{Formula for regression}

\item{se}{Standard error column name}
}
\value{
Meta-regression results
}
\description{
Meta-Regression Analysis
}
