% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{plot_bayesian_diagnostics}
\alias{plot_bayesian_diagnostics}
\title{Plot Bayesian Diagnostics}
\usage{
plot_bayesian_diagnostics(x)
}
\arguments{
\item{x}{cbamm_bayesian object}
}
\value{
A grid of diagnostic plots
}
\description{
Plot Bayesian Diagnostics
}
