% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{prediction_interval}
\alias{prediction_interval}
\title{Prediction Interval}
\usage{
prediction_interval(estimate, se, tau2, df)
}
\arguments{
\item{estimate}{Overall effect estimate}

\item{se}{Standard error}

\item{tau2}{Between-study variance}

\item{df}{Degrees of freedom}
}
\value{
Prediction interval
}
\description{
Prediction Interval
}
