% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{print.cbamm}
\alias{print.cbamm}
\title{Print CBAMM Results}
\usage{
\method{print}{cbamm}(x, digits = 4, ...)
}
\arguments{
\item{x}{CBAMM object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments}
}
\description{
Print CBAMM Results
}
