% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{print.cbamm_bayesian}
\alias{print.cbamm_bayesian}
\title{Print Bayesian CBAMM Results}
\usage{
\method{print}{cbamm_bayesian}(x, digits = 3, ...)
}
\arguments{
\item{x}{cbamm_bayesian object}

\item{digits}{Number of digits to display}

\item{...}{Additional arguments}
}
\description{
Print Bayesian CBAMM Results
}
