% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{print.cbamm_cumulative}
\alias{print.cbamm_cumulative}
\title{Print Cumulative Results}
\usage{
\method{print}{cbamm_cumulative}(x, ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Print Cumulative Results
}
