% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{print.cbamm_metareg}
\alias{print.cbamm_metareg}
\title{Print Meta-regression Results}
\usage{
\method{print}{cbamm_metareg}(x, digits = 3, ...)
}
\arguments{
\item{x}{cbamm_metareg object}

\item{digits}{Number of digits}

\item{...}{Additional arguments}
}
\description{
Print Meta-regression Results
}
