% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{print.cbamm_subgroup}
\alias{print.cbamm_subgroup}
\title{Print Subgroup Results}
\usage{
\method{print}{cbamm_subgroup}(x, ...)
}
\arguments{
\item{x}{cbamm_subgroup object}

\item{...}{Additional arguments}
}
\description{
Print Subgroup Results
}
