% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{robust_rma}
\alias{robust_rma}
\title{Robust Meta-Analysis Wrapper}
\usage{
robust_rma(yi, sei, method = "REML", ...)
}
\arguments{
\item{yi}{Effect sizes}

\item{sei}{Standard errors}

\item{method}{Estimation method}

\item{...}{Additional arguments}
}
\value{
Meta-analysis result
}
\description{
Robust Meta-Analysis Wrapper
}
