% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{safe_divide}
\alias{safe_divide}
\title{Safe Division}
\usage{
safe_divide(x, y, na_value = NA)
}
\arguments{
\item{x}{Numerator}

\item{y}{Denominator}

\item{na_value}{Value to return when dividing by zero}
}
\value{
Result of division
}
\description{
Safe Division
}
