% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{se_to_var}
\alias{se_to_var}
\title{Convert Standard Error to Variance}
\usage{
se_to_var(se)
}
\arguments{
\item{se}{Standard error}
}
\value{
Variance
}
\description{
Convert Standard Error to Variance
}
