% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{standardize_cbamm_data}
\alias{standardize_cbamm_data}
\title{Standardize CBAMM Data}
\usage{
standardize_cbamm_data(data, verbose = FALSE)
}
\arguments{
\item{data}{Data frame to standardize}

\item{verbose}{Logical for verbose output}
}
\value{
Standardized data frame
}
\description{
Standardize CBAMM Data
}
