% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{subgroup_analysis}
\alias{subgroup_analysis}
\title{Subgroup Meta-Analysis}
\usage{
subgroup_analysis(data, subgroup, yi = "yi", se = "se")
}
\arguments{
\item{data}{Data frame with effect sizes}

\item{subgroup}{Subgroup variable name}

\item{yi}{Effect size column name}

\item{se}{Standard error column name}
}
\value{
Subgroup analysis results
}
\description{
Subgroup Meta-Analysis
}
