% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{summary.cbamm}
\alias{summary.cbamm}
\title{Summary CBAMM Results}
\usage{
\method{summary}{cbamm}(object, ...)
}
\arguments{
\item{object}{CBAMM object}

\item{...}{Additional arguments}
}
\description{
Summary CBAMM Results
}
