% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{validate_cbamm_data}
\alias{validate_cbamm_data}
\title{Validate CBAMM Data}
\usage{
validate_cbamm_data(data)
}
\arguments{
\item{data}{Data frame to validate}
}
\value{
Validated data frame with yi and vi columns
}
\description{
Validate CBAMM Data
}
