% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-main.R
\name{var_to_se}
\alias{var_to_se}
\title{Convert Variance to Standard Error}
\usage{
var_to_se(var)
}
\arguments{
\item{var}{Variance}
}
\value{
Standard error
}
\description{
Convert Variance to Standard Error
}
